python: JobInfo: set default for job priority
Problem: Job priority is not assigned to jobs until after the
PRIORITY, so the 'priority' attr may be missing from the job-info
response for jobs in DEPEND or PRIORITY state. This can lead to
an error from flux-jobs(1) when the priority is requested.

Add a default value for priority of the empty string when a priority
value for jobs is not available.

Fixes #3475