github: only run unit tests in asan builder
Problem: The asan build takes a very long time to run in the Github
actions runner, and there have been numerous cases of timeouts and
other instability when running the entirety of `make check` with
address santizer enable.

While the hangs are under investiation, run only the unit tests
in the asan builder. This should allow PRs to be merged efficiently,
but with some protection from obvious errors which would be caught
by the address sanitizer.