python: add set_treedict(), TreedictAction to flux.util
Problem: The set_treedict() function is generally useful, but is
currently embedded as a private method of flux.job.Jobsepec.

Move the set_treedict() implementation to flux.util so it may be used
by other callers.

Also add a TreedictAction class for use with utilities that take a
series of key=value "treedict" like args and want a dictionary in return.