testsuite: use rc1 scripts to set resources
Problem: In many job-manager tests, resources are manually set up
and configured.  In many cases, this is unnecessary, as the rc1-job
run control script will setup resources for the tests if the
correct configuration is done when calling test_under_flux().

Solution: When possible, configure resources by passing the
correct number of ranks and/or setting TEST_UNDER_FLUX_CORES_PER_RANK
when calling test_under_flux() and using the job personality.

Update test description to be more clear about the resources
that are being configured.