testsuite: add TEST_UNDER_FLUX_NO_JOB_EXEC option
Problem: In many job-manager test files, we remove the job-exec
module at the start, to avoid running jobs in the tests.  This is
repeated in many files.

Solution: Add an environment variable TEST_UNDER_FLUX_NO_JOB_EXEC that
will not load job-exec if it is set.  Set this environment variable
appropriately and remove the new unnecessary job-exec module removes.