testsuite: re-order tests to remove queue commands
Problem: In a number of job-manager tests, the sched-simple
scheduler is unloaded, jobs are submitted, the queue is stopped,
the sched-simple scheduler is reloaded, and then the queue is
restarted.  Often the queue stop/start is unnecessary if we re-order
how the tests are run.

Solution: Instead of re-loading the sched-simple module after
submitting jobs, re-load the sched-simple module before submitting
jobs.  This allows us to remove the stop/start queue in many circumstances.
Only in the few cases where necessary, stop/start the queue.