job-manager: submit: split submit routines to prep for validation
Problem: The job-manager.submit handler adds jobs to the active
job hash at the same time as decoding the "batch" payload. This
restricts the flexibility of the code and, for example, prevents
easily adding code to validate each job before it is added to the
active job hash.

Split the submit functions into two distinct operations:

 1. submit_jobs_to_list(): decode the submit payload into a
    "newjobs" zlistx_t. A zlistx_t is used here to allow items to
    be deleted (e.g. invalid or already hashed jobs) safely while
    iterating the list.

 2. submit_hash_jobs(): Add the jobs to the active_jobs hash.

With this approach the jobs in the newjobs list could be validated
before they are enqueued in the active_jobs hash. For now, however,
there should be no functional difference.