job-ingest: support per-job errors from job-ingest.submit
Problem: When job-ingest announces a set of new jobids to to the
job-manager, it is an all or nothing situation. Either the entire
batch succeseds or fails. In the future, the job-manager may want
to reject individual jobs (e.g. for user-based limits), and so may
want to return per-job errors.

Extend the processing of job-manager.submit reponses to include an
optional "errors" array in the payload. The errors array will include
a set of [jobid, errmsg] tuples for each failed job.