job-manager: allow plugins to validate jobs
Problem: job-manager plugins may wish to enforce policy by rejecting
jobs, but currently this can only be done by raising an exception and
canceling submitted jobs. This is not ideal since jobs will appear to
have been successfully submitted by a user, only to have the job fail
at a later time.

Introduce a 'job.validate' callback for jobtap plugins which is called
before jobs are inserted into the job manager's active job hash. If a
plugin returns < 0 from this callback, then the job will be considered
to be rejected and will be removed from the list of jobs targeted for
submission.

Rejected jobs, along with an error message optionally provided by the
plugin, will be noted in the response to the job-ingest module, which
will in turn forward the failure on the original submitter.