job-manager: move job.new callback to just after validation
Problem: The job.new jobtap callback is not invoked until after all
jobs in a batch have been validated. This could lead to an inability
to enforce limits. For example, if a batch of jobs puts a user over
a limit halfway through the batch, a plugin which uses job.new to
implement counters may have their counters incremented too late.

Move the job.new callback to submit_validate_job(), so that any
accounting done by a plugin in job.new will be complete before the
next job is validated. The job.new callback is only invoked for
jobs that are not already in the active jobs hash.