job-manager: add flux_jobtap_reject_job()
Problem: Rejecting a job from a jobtap job.validate hook and including
an appropriate error message is a multi-step process. The error message
has to be added to the plugin "out" arguments under the proper name, and
the function has to return < 0.

Encapsulate job validation failure in a single new call,

  int flux_jobtap_reject_job (flux_plugin_t *p,
                              flux_plugin_arg_t *args,
                              const char *fmt, ...);

which sets the appropriate error message on the plugin out arguments.
This function always returns -1 so that it can be used on exit from
the callback as:

 return flux_jobtap_reject_job (p, args, "Too many submitted jobs");