job-ingest: refactor context initialization
Problem: The job-ingest module initialization of internal context
object is handled directly in mod_main(). This reduces flexibility
especially as it pertains to processing the module command line
options.

Move much of the ctx initialization into a new job_ingest_ctx_init()
function. Process module arguments first so that the module fails
as soon as practical when invalid options are provided.