job-ingest: add batch-count=N module option for testing
Problem: It is not easy to force the job-ingest module to include
multiple jobs in a batch during testing since the batching is
done on a timer basis.

Add a batch-count=N module option which submits a batch based on the
number of jobs in the batch instead of a timer. This should only be
used for testing since you must ensure at least N jobs are submitted
or the submissions will appear to hang.