heartbeat: add new module
Problem: heartbeat generation is a feature built into the broker,
but in some ways, built-in services hinder broker development
and make them harder to understand because they are not isolated.

Re-implement heartbeat generation in a loadable module.

The heartbeat period can be set as a module option period=FSD.
Any valid FSD is accepted, whereas previously the period was limited
to the range of 0.01s to 30s.

The heartbeat period can be fetched with the heartbeat.get RPC for
testing or whatever.