flux-mini: allow Xcmd to be used in place of argparse args
Problem: When applying Xcmd (eXtensible Cmd) modified arguments
to the argparse `args` during multiple submissions. The `args`
object is overwritten. When args differ for each submitted job,
the callbacks specific to that job therefore do not have a valid
copy of their args, but in fact are referencing a global copy
which is being or has been changed.

Add a getattr method to the Xcmd object which falls back to the
original args used to create the new args. This allows the original
argparse args to become read-only, and only changes to those args
are saved in the Xmcd object. The Xmcd object may then be used in
place of args throughout, giving each submitted job and its callbacks
access to a private copy of args that will not change.