flux-mini: allow empty inputs in Xmcd constructor
Problem: Passing an empty `inputs` list to Xmcd() requires supplying
an empty list in args. This seems unnecessary.

Allow empty inputs by setting a default of inputs=None. If inputs is
not set then an empty list is used.

This allows Xcmd() to be used to substitute only keyword args.