flux-mini: submit: add --log and --log-stderr options
Problem: When submitting many jobs with --cc it is convenient to
capture flux-mini output in separate per-job files, especially for
testing purposes. However, submit/bulksubmit always emits jobids
(and output with --watch) to the terminal.

Add options --log=FILE and --log-stderr=FILE which allow submit
and bulksubmit output to be redirected to one or more logfiles.
The command is most useful when it includes "{cc}" or other
substitutions which would make a unique per-job logfile.

For example, to capture the jobids of many submitted jobs in separate
files, simply use:

 flux mini submit --cc=0-100 --log=job{cc}.id hostname

and jobids will be logged to files job0.id, job1.id, ...