flux-mini: bulksubmit: add {seq1} keyword arg
Problem: The {seq} keyword is replaced by the current overall sequence
in the bulksubmit generation of commands, but it starts at 0, and some
users may enjoy to start their sequences at 1 instead.

Attempt to make everyone happy by adding a new {seq1} keyword that will
be substituted with the sequence number starting at 1.