lua: work around 32b int conversion errors in Lua <= 5.2
Problem: Previous to Lua 5.3 64 bit integers are not supported on 32
bit systems.  This is in part due to the fact that there is no internal
integer type in Lua (all numbers are stored as double precision
floating-point values), and also because the lua_pushinteger()
and lua_tointeger() calls use the type `ptrdiff_t` which is 32
bits on a 32b system. This causes a problem when decoding/encoding
payloads with flux_jobid_t. On i386 the conversion from ptrdiff_t
to uint64_t results in an incorrect conversion, and as a result the
exec-service.lua script used in t2401-job-exec-hello.t responds to
the exec.start request with an invalid jobid, resulting in a test hang.

To workaround the issue, convert JSON integers directly to double
when decoding if the Lua version is < 5.3 *and* the size of ptrdiff_t
is 4 bytes. The integer values will be stored as double internally
anyway. Also, when encoding a Lua integer to JSON, use lua_tonumber()
instead of lua_tointeger(). This will prevent most type conversion
errors.

In Lua 5.3 the lua_Integer type is 64 bits, so we can fall back to
more sane behavior on this version.