docker: add --platform option to docker-run-checks
Problem: There is no way to pass the --platform request to the underlying
docker build and run commands used in docker-run-checks.sh when they are
supported. This limits the ability to run docker-based checks on alternate
architecture types, e.g. linux/386.

Add a --platform option to docker-run-checks.sh that, if set, is passed
along to the underlying docker commands. Additionally, check for buildx
support and use "docker buildx build" if available since it will have
necessary support for multiple architectures.