github: add a 32 bit bionic build to ci
Add a 32 bit bionic build by adding --platform=linux/386 to
generate_matrix.py. This build also includes the new 'needs_buildx'
parameter so that the buildx extension for docker is pulled into
the builder.

Note: this is not an actual qemu i686 build, but a build on x86_64
with a 32bit runtime, because the qemu build was too slow. However,
even building in the 32bit runtime should uncover most of the integer
conversion errors that are commonly the cause of breakage on this
platform.