github: add docker buildx if necessary
Problem: Use of the docker-run-checks.sh --platform option may require
that 'docker buildx' be installed in the Github runner instance, but we
don't want to install buildx when it isn't necessary.

Add a step to the ci github action to pull in buildx if 'needs_buildx"
is set in the matrix.