libflux: add flux_sync_create()
Problem: heartbeat timing is configurable and can be unpredictable,
yet there is no simple way to ensure that a heartbeat callback is
called with a minimum frequency.

Add flux_sync_create(), which creates a future that is fulfilled each
time the heartbeat message is received.  Subscribe and unsubscribe are
handled internally.

A deadline can be established using existing future timeouts,
and the future can be used in synchronous and reactive contexts
like any flux future.

Malformed heartbeat events, or events not sent by the instance
owner are ignored and do not fulfill the future with an error.

If desired, a minimum time between fulfillments can be set, to
protect from thrashing on expensive activities when several heartbeats
are "bunched up" due to network congestion and arrive in quick succession.

Add unit test.