job-exec: fix segfault in error callback
Problem: In job-exec/exec.c:error_cb() the return code of
flux_subprocess_get_cmd() is not checked, but this function may
return NULL if there was a failure in creation of the flux_cmd_t
object for the flux_subprocess_t. This causes a segfault when
handling errors for many cases, for example, when the broker has
exceeded the max number of open files.

Check the return code from flux_subprocess_get_cmd() to avoid
the NULL dereference. Exclude the command arg0 from the exception
message when it is not available.

Fixes #3505