flux-mini: fix counting error with --watch --progress
Problem: The job "finish" event is always posted when there is an
"alloc" event, but the code for progress_update() in flux-mini submit
assumes that the "finish" event is instead  associated with  the
"start" event. This leads to a counting error when a job gets an
exception after alloc but before start, since the job is not yet
marked as running.

Set the internal state of a job to 'running' and increment the running
jobs count at the "alloc" event instead of the "start" event. This
fixes the counting error since every finish event can decrement the
running job counter.