flux-mini: fix running job accounting with --progress
Problem: When processing the "alloc" event in flux-mini submit --wait,
a test for "args.watch" was inadvertently required before updating the
job state, thereby leaving the job state as "submit", which results in
incorrect accounting in the --progress display.

Remove the test for args.watch when processing the "alloc" event in
event_watch_cb().