job-exec: ensure jobs with exceptions have status != 0
Problem: If a job gets an exception before it has started (unlikely),
then the job will still finish with status == 0 since the job releases
resources and finishes without reaping any job shells which would
adjust the job->wait_status reported in the finish event. Tools
which use the finish event status to determine if a job failed will
incorrectly assume these types of jobs have succeeded.

Check for a job exception when posting finish event back to the
job-manager, and ensure the finish status will be reported as nonzero.