testsuite: fix unreliable bulksubmit tests
Problem: The flux-mini bulksubmit --progress tests in
t2703-mini-bulksubmit.t assume that all jobs will be pending before
any job is running or complete. This is not true since jobs could be
started and even complete before all jobs have been fully submitted
(the time at which they are marked pending).

Fix the test to ensure that a *known* initial condition for the
progress display is emitted, namely that a single job is pending
while all other counters are 0.