librouter: pause usock server acceptor on EMFILE
Problem: When server_accept() in librouter fails to accept a new
connection due to exceeding the max number of open file descriptors,
it just returns immediately and the calling process goes back to
sleep in the reactor. However, when accept() fails due to exceeding
the max number of file descriptors (EMFILE or ENFILE), the socket is
immediately ready and the accept() is tried again, fails, etc, until
something else in the broker frees a file descriptor and the accept()
can succeed. This causes a busy loop which, especially int case of
the connector-local module, can starve out other messages and slow
down resolution of the issue.

When server_accept() fails with ENFILE or EMFILE, "pause" the watcher
on the server file descriptor for 10ms to avoid this condition. This
limits the rate at which accept(2) is retried, and allows other parts
of the system to proceed and (hopefully) free a file descriptor.