broker: remove prep/check generated keepalives
Problem: The keepalive messages generated for all modules in the
prep and check watchers may have a significant impact on a busy
Flux instance, since modules with a large queue of messages will
be entering and exiting the reactor at a high rate. Since this
results in 2 messages to the broker for every 1 unit of work the
module may do, this could significantly impact the broker's ability
to handle other messages in the system.

Remove the keepalive generated by the check watcher completely, and
only keep the prepare watcher so that a keepalive message can be
sent to the broker one time at module initialization to let the
broker know the module is entering the reactor and the module is
ready in the case the module does not call `flux_module_set_running(3)`
manually.

Fixes #3515