libflux/future: use recurring timer
Problem: when a continuation is registered with a timeout,
and the future is fulfilled multiple times, the timeout is
stopped and started for each fulfillment.  This is not efficient.

A more efficient method is to create a recurring timer and
call flux_timer_watcher_again() on it each time the future is
reset.  This avoids processing associated with stopping and
starting the watcher (removing and re-adding it to the event loop)
as described here:

http://pod.tst.eu/http://cvs.schmorp.de/libev/ev.pod#code_ev_timer_code_relative_and_opti

(We were using method 1; this transitions to method 2.)