etc: Load job-info on all ranks
Problem: The job-info module was previously only loaded
on rank 0, as the job listing service could take up a significant
amount of memory across an allocation.  Now that the job listing
service is no longer a part of job-info and is in the job-list module,
that issue no longer exists.  By loading the job-info module on
all ranks, there can be performance improvements.

Solution: Update all rc scripts to load the job-info module
on all ranks.