job-ingest: use flux-job-validator as default validator
Problem: The job-ingest module requires a validator path be set,
either by built-in configuration or on the command line, but a new
flux job-validator command is available which replaces the use of
multiple job validator scripts.

Switch the job-ingest module to use `flux job-validator` as the
built-in default validator. Add a new `validator-plugins=LIST` option
to simplify setting the list of validator plugins to be used in job
validation and enabling the use of multiple simultaneous validators.

Drop the internal configuration for default validator path and
arguments.  This is no longer necessary.

Keep the `validator=PATH` option for now to avoid breaking existing
tests.