github: explicitly checkout tag if creating a release
Problem: When a tag is pushed, GitHub Actions does not appear to
properly checkout the tag such that `git describe` returns a SHA
suffix version instead of the version being tagged.

When creating a release, add a step after git clone that explicitly
checks out the tag ref to hopefully work around this problem.