sched-simple: add sched.feasibility RPC
Problem: Jobs that are not feasible or are otherwise invalid
specifically for sched-simple are not rejected at ingest, but instead
sit in the queue and have a job exception raised when the scheduler
attempts to allocate resources on the job. It would improve usability
to reject infeasible jobs on ingest.

Add a sched.feasibility RPC which, given a jobspec, will return an
error if the job is infeasible or is otherwise invalid for scheduling
with sched-simple (e.g. includes resources other than cores and nodes)

This RPC can then be used with the job validator to ensure jobs are
feasible at job ingest, returning a useful error message instead of
queueing an invalid job.