python: add scheduler feasibility validator plugin
Problem: There is no way to reject jobs at ingest which are not
feasible.  Instead, they are queued and later have a job exception
raised when the scheduler considers them for an allocation.

Add a flux.job.validator plugin that uses the sched.feasibility RPC
to ensure job feasibility (or anything else the scheduler might want
to check) at ingest. This allows the job to be rejected before being
ingested, improving usability.

Fixes #269
Fixes #2049