testsuite: run validator tests in parallel
Problem: Running the test_valid() and test_invalid() jobspec validation
tests serially slows down the test.

Use bulksubmit to submit the tests as jobs so that they run in
parallel.

The queue must be running for our test jobs to execute, so do not run
`flux queue stop` until later in the testsuite.  Be sure to submit jobs
for validation with urgency == 0 so that the jobs do not actually run.