broker: drop broker.rundir and overlay init callback
Problem: the creation of 'broker.rundir', a directory that
is guaranteed unique for each broker, as opposed to 'rundir',
which is not always unique, is no longer necessary.

Change the local URI from local://<rundir>/<rank>/local to
local://<rundir>/local-<rank>.

Drop the init callback from overlay.c that was needed to create
the broker.rundir between discovering the rank and creating an
ipc:// tbon endoint in it when bootstrapping with PMI.

Update flux-broker-attributes(7).
Update overlay unit test.