modules/kvs: ensure cache entries survive kvs tnxs
Problem: In the kvstxn internal API, references to cache
entries are internally stored and expected to be available
to callers.  However, under an extremely busy system, those
cache entries could expire before users can gain access to the
cache entries.

Solution: Take a reference on cache entries that the kvstxn
library has internal references to to ensure they don't
expire from the cache until the user is done with the transaction.

Fixes #3549