libflux: add flux_msg_match_route_first()
Problem: flux_msg_get_route_first() copies its route result
to a heap-allocated buffer and requires the user to free it;
this is inconvenient when performing the common task of matching
a disconnect message to a list of pending requests.

Add flux_msg_match_route_first() which returns true if
two messages have the same first route (the request sender),
and false if they have different senders, no routes, or are NULL.
The routes are compared in place without allocating memory.

Update unit test.