flux-job: add exec subcommand
Problem: the python flux mini command is a bit slow to start
up for some use cases.

Add a C job submission front end that is a little faster, and
uses the job-manager.runjob RPC.  Options are very stripped down.

Job execution occurs in one RPC round trip.  Once the job is complete,
output is fetched with another RPC round trip.

Since job-manager.runjob is not available to guests, and output is
fetched directly from the KVS rather than through the job-info proxy,
this command is only available to the instance owner.