job-manager: iterate active jobs safely in jobtap
Problem: When a jobtap plugin is loaded, all active jobs are iterated
using zhashx_first/zhashx_next and the plugins `job.new` callback is
invoked on each job. However, this is not safe because the plugin could
call a function which results in a zhashx_lookup(), which apparently
resets the zhashx cursor.

Iterate jobs safely by using zhashx_values() to create a local list
of jobs which won't be modified by any plugin action.