job-manager: jobtap: notify loaded plugin of jobs in DEPEND
Problem: If a jobtap plugin is loaded when jobs are in the DEPEND
state, then there is no way for the plugin's job.state.depend callback
to be called, so the job may get stuck in this state.

Call the job.state.depend callback for jobs in the DEPEND state during
plugin load. This allows plugins that handle dependencies to initialize
whatever state is necessary to manage a potential job dependency.

Note that this really only works for the case of reloading the same
dependency plugin, since dependency references are handled by name.
A dependency added by a different dependency plugin may cause a job
to be stuck in the DEPEND state until that dependency is removed.