job-manager/journal: use flux_msglist
Problem: job-manager/journal implements one-off disconnect, cancel
handlers.

Instead of placing request message and their context in a struct
and keeping a list of the structs, place the context struct in the
message's aux hash.  This reduces code complexity since the context
is destroyed with the message.

Then track the pending requests in a flux_msglist, which allows the
disconnect helpers to be used.