resource/acquire: use flux_msglist
Problem: resource/acquire uses one-off code to handle disconnect
and cancel requests.

Instead of placing request message and context in a struct
and keeping a list of the structs, place the struct in the
objects in the message's aux hash.  This reduces code complexity
since the context is destroyed with the message.

Then track the pending requests in a flux_msglist, which allows the
disconnect helpers to be used.