libjob/specutil: add jobspec helpers
Problem: we need to maniuplate a bit of RFC 14 jobspec
in C to implement 'flux job exec'.

Add some helpers to libjob that are for flux-core internal use.
This is not intended to be a generalized C jobspec library,
but putting it in a library makes it easier to develop standalone
tests and move functionality back and forth between the job manager
and flux-job as needed during development.

Add unit tests.