python: support dependencies attr in JobInfo/JobList
Problem: The list of job dependencies is available from the job-list
module, but the Python bindings and flux-jobs(1) do not yet support
fetching this attribute.

Support the "dependencies" attribute in the Python JobInfo and JobList
interfaces, and add the attribute to flux-jobs(1) so it can be queried
with `flux jobs -o {dependencies}`.