libutil: add grudge set implementation
Problem: Proposed management of job dependencies via dependency-add
and dependency-remove events in the job-manager and job-list modules
requires a modified Set implementation which supports add-only-once
semantics. That is, values removed from the set are "retired" and
cannot be added again.  Since the code will be shared between modules
that track the job dependency state, we need a shared class with these
semantics.

Add a very small new class to libutil called a "grudge set". A grudge
set operates as described above -- it is a set that allows values to
only ever be added once.

This will be used to implement job dependency tracking in Flux modules.