libflux: add flux_future_fulfill_next(3)
Problem: When using chained composite futures
(i.e. flux_future_and_then(3)), it is sometimes convenient to be able
to fulfill the "next" future in the chain immediately with a custom
result. For example, if using flux_future_and_then(3) to allow some
processing of the result of a different future-based call and return
the processed result to the user in a different future. However,
the current paradigm of using flux_future_continue(3) does not easily
allow this, especially in the case of futures which may have multiple
results pending.

Add new function to the API, flux_future_fulfill_next(3), which is
identical to flux_future_fulfill(3), but fulfills the cf->next future
in a chained future, instead of the future itself.