libkvs: save aux data in future instead of handle
Problem: In flux_kvs_copy() and flux_kvs_move() some auxiliary data
was stored in the flux handle, with a destructor set for when the handle
is destroyed.  However, the flux handle is a long lived object, typically
staying alive for the length of a flux instance.  Therefore, the memory
is basically never freed during the existence of the instance.

This continual memory allocation led to a very large auxiliary list
in the underlying implementation and poor performance for long running
instances with many jobs.

Solution: Move the auxiliary data from the flux handle to the futures
used within flux_kvs_copy() and flux_kvs_move().  These futures are
short lived, therefore auxiliary memory will be cleaned up more
regularly.