python: add flux.job.result() and flux.job.result_async()
Problem: There is no equivalent to flux.job.wait() for non-waitable
jobs in the Python bindings.

Add a binding for flux_job_result(3), which is similar to flux_job_wait(3),
but processes the job eventlog and returns either the raw "result" dict
(get_dict()) or a JobInfo object instantianted from this dict (get_info()).