libflux/handle: drop paranoid matchtag free code
Problem: flux_matchtag_free() scans the handle "backlog" of
received messages for responses containing the matchtag that
is proposed to be freed, and removes any matches, but the RPC
code knows not to call flux_matchtag_free() if the matchtag
cannot be safely retired.

Drop this code, which adds overhead for little benefit.