content-cache: store requests in a stack
Problem: each content-cache hash entry may have requests attached
to it that are answered when the entry changes state.  The flux_msglist,
which is based on zlistx_t has a higher memory overhead than necessary
for this application.

Implement a simple message stack that uses less memory.

Memory cost for stack:
  0 bytes fixed overhead
  16 bytes per message (msg and next pointers)

Memory cost for flux_msglist:
  60 bytes fixed overhead (incl. 44 bytes for zlistx_t)
  28 bytes per message (zlistx_t node_t)