content-cache: convert log to assertion
Problem: when a content cache entry is removed, an LOG_ERR
message is generated if it has pending load/store requests
attached.  This should not happen, and if it does, it should
be noticed.

Convert log message to assertion.

Drop flux_t handle from cache entry since it is no longer used.