content-cache: use LRU to purge
Problem: the periodic content-cache purge is costly due to the
fact that entries are scanned in random order.

Keep cache entries in a LRU list.  Entries are added to the front
of the list.  Every time an entry is looked up, it is moved to
the front of the list.

When the periodic sync callback is called, scan the LRU list starting
with least recently used, removing entries that are not busy (e.g. w/o
pending load/store requests).  Terminate the scan when:

- the total cached data <= 'content.purge-target-size' bytes.

- the newer entries in the LRU are reached, where "newer" is defined
  as as used within 'content.purge-old-entry' seconds.

The 'content.purge-large-entry' tunable is removed.  Singling out the
larger entries would require a separate list or a complete scan,
and this doesn't seem worth it.  To compensate for this, the
'content.purge-target-entries' tunable is also removed, and we only
suspend purging based on the overall size, not the number of entries.