flux-job: add -j, --json option to flux-job status
Problem: It may be useful to dump the underlying JSON object obtained
for jobs with flux_job_result(3) when using flux-job status, but this
is currently not possible.

Add a new option, -j, --json, to `flux-job status` which will dump the
JSON object used to fulfill the future returned by flux_job_result(3).