overlay: check configuration of downstream peers
Problem 1: if a node in the system instance has the wrong version
of flux-core installed, mismatched versions could try to interoperate.

Problem 2: if a broker in the system instance has an incorrect
bootstrap configuration, it could try to connect to the wrong
TBON parent.

Add an overlay.hello handshake initiated by downstream TBON peers.
If the handshake fails, print a human-readable error message on stderr
and post 'parent-fail' to the state machine, to trigger broker exit
with a nonzero exit code.  In a system instance, systemd should capture
the error in the journal, and not automatically restart the broker.

If a peer tries to say hello, but the identity pushed onto the route
stack by the ROUTER socket does not match an expected child rank
in the current geometry, respond with an error.

If a peer tries to say hello with an incompatible flux-core version,
respond with an error.  For now, require major, minor, and, patch
versions to agree.  Later we can relax this for versions that we intend
to interoperate.

New overlay interfaces:

overlay_parent_error()
  call from overlay monitor callback to determine if handshake
  with parent has completed unsuccessfully

overlay_parent_success()
  call from overlay monitor callback to determine if handshake
  with parent has completed successfully - currently only used in test

overlay_set_version()
  (for testing) override the comiled-in version, to simulate
  overlay connection from an incompatible flux-core version

overlay_set_rank()
  (for testing) override the rank set by overlay_set_geometry(),
  to simulate overlay connection from non-peer

Update overlay unit test.

Fixes #3230