broker: isolate broker UUID in overlay.c
Problem: the broker assumes that its ZMQ_ROUTING_ID is its rank,
but that identity is the responsibility of the overlay class.

Specifically, the rank is assumed
1) when determining whether the next hop in a response route stack
   is a broker peer or a local broker module
2) when pushing its identity onto a request message sent to
   a broker module over a ZMQ_PAIR socket
3) when appending the broker ID to a ping response payload

Add overlay accessors:
  overlay_get_uuid()
  overlay_uuid_is_parent()
  overlay_uuid_is_child()

And use those in the above circumstances.

It should now be possible to change from rank to an actual UUID
without breaking anything outside of overlay.c.