broker: set PMI_process_mapping when size=1
Problem: mvapich2-2.3 fails with an "invalid tag" error from
MPI_Init() when launched from a single broker that was started
with flux-start.

Apparently mvapich becomes sad when the 'PMI_process_mapping' key
is not found in the PMI KVS.  Other MPIs treat the missing key as
optional, so Flux does not create it in certain situations:
1. when the enclosing instance did not provide a mapping to Flux
2. when multiple brokers per node are launched on mutiple nodes
3. when the mapping exceeds the PMI-1 fixed max KVS value length

A workaround for mvapich is to run the job with -opmi.clique=pershell.
However, that is an inappropriate default, as it causes openmpi to
segfault in some circumstances (#3551).

This commit partially addresses the problem by ensuring that
PMI_process_mapping is created when the broker is launched as a
singleton outside of a PMI environment, such as when run by
flux start without arguments, or when the broker finds slurm's
libpmi.so and is not running in a slurm job.

Fixes #3592