job-list: Support inactive-cache-size config
Problem: Presently the job-list module caches a large amount of
job information for all jobs, including those that are inactive.
Over a long period of time, this cache of information grows and
uses a non-trivial amount of memory.  This leads to poor performance
over time.

Solution: For users that wish for superior performance, a job-list
configuration called "inactive-cache-size" is now supported.  It
will limit the number of inactive jobs that will be cached by the
job-list module, therefore limiting the amount of memory that will
be used over time.

This option will limit the number of inactive jobs listed by the
module, therefore affecting job listing tools such as flux-jobs.

Fixes #3607